essaie

+++ 21meeseeks-ejb/src/main/java/services/ResourceService.java
@@ -1,17 +1,26 @@
 package services;
 
 import java.util.List;
+import java.util.Set;
 
 import javax.ejb.LocalBean;
 
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
+import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
+import javax.persistence.Query;
 import javax.persistence.TypedQuery;
 
 import interfaces.ResourceServiceLocal;
 import interfaces.ResourceServiceRemote;
+import entities.DayOff;
+import entities.Note;
 import entities.Resource;
+import entities.Resume;
+import entities.Seniority;
+import enums.Availability;
+import enums.ContractType;
 
 /**
  * Session Bean implementation class ResourceService
@@ -30,7 +39,33 @@ public class ResourceService implements ResourceServiceRemote, ResourceServiceLo
     }
 	@Override
 	public int ajoutRessource(Resource r) {
+		
+		r.setResume(em.find(Resume.class, r.getResume().getIdResume()));
+		r.setSeniority(em.find(Seniority.class, r.getSeniority().getIdSeniority()));
+		//em.merge(r);
+		/*
+		Set<Note> noteliste = r.getNotes();
+		Note note = em.find(Note.class, r.getNotes().iterator().next().getIdNote());
+		noteliste.add(note);
+		r.setNotes(noteliste);
+		System.out.println(r.getNotes());
+		System.out.println(note);
+		
+		
+		DayOff dayOff = em.find(DayOff.class, r.getDayOffs().iterator().next().getIdLeave());
+		Set<DayOff> dayOffliste = r.getDayOffs();
+		dayOffliste.add(dayOff);
+		r.setDayOffs(dayOffliste);
+		System.out.println(r.getDayOffs());
+		System.out.println(dayOffliste);
+		*/
+		
+		//em.merge(r);
+		
+		System.out.println(r);
+		
 		em.persist(r);
+		
 		return r.getIdUser();
 	}
 	@Override
@@ -41,8 +76,31 @@ public class ResourceService implements ResourceServiceRemote, ResourceServiceLo
 	@Override
 	public Boolean deleteResource(int id) {
 		Resource r=em.find(Resource.class, id);
-		em.remove(r);
-		return true;
+		try{
+		if(r != null){
+			Query query = em.createQuery("DELETE Level l WHERE l.resources.idUser=:resources");
+			query.setParameter("resources", r.getIdUser()).executeUpdate();
+			
+			Resume query2 = (Resume)em.createQuery("SELECT r FROM Resume r , Resource rs WHERE rs.idUser = :resourceId and r.idResume = rs.resume").setParameter("resourceId", r.getIdUser()).getSingleResult();
+			Query query3 = em.createQuery("UPDATE Resource r SET r.resume = null WHERE r.idUser = :id");
+			query3.setParameter("id", r.getIdUser()).executeUpdate();
+			Query query4 = em.createQuery("UPDATE Resource s SET s.seniority = null WHERE s.idUser = :id");
+			query4.setParameter("id", r.getIdUser()).executeUpdate();
+			r.setSeniority(null);
+			
+			em.remove(query2);
+			//System.out.println(level);
+			//em.remove(level);
+			em.remove(r);
+			
+			return true ;
+		}
+		}catch(NoResultException e){
+			em.remove(r);
+			return true ;
+		}
+		
+		return false;
 	}
 	@Override
 	public void updateResource(Resource r) {
@@ -57,5 +115,30 @@ public class ResourceService implements ResourceServiceRemote, ResourceServiceLo
 		List<Resource> listResource = q.getResultList();
 		return listResource;
 	}
-
+	@Override
+	public int lastIndex() {
+		try {
+			TypedQuery<Integer> index = em
+					.createQuery("select max(r.idUser) from Resource r", Integer.class);
+			
+			return index.getSingleResult();
+		} catch (NoResultException e) {
+			System.out.println(e);
+			return 0;
+		}
+	}
+	@Override
+	public List<Resource> getResourceByType(ContractType type) {
+		return em.createQuery("select r from Resource r where "
+				+ "r.contractType=:type", Resource.class)
+				.setParameter("type", type).getResultList();
+	}
+	@Override
+	public List<Resource> getResourceByAvaibility(Availability avaib) {
+		return em.createQuery("select r from Resource r where "
+				+ "r.availability=:avaib", Resource.class)
+				.setParameter("avaib", avaib).getResultList();
+	}
+	
+	
 }
